examples/volatile_query_detection/requirements.txt
psycopg2-binary>=2.9
sentence-transformers>=2.0
requests>=2.28
